requirements.txt
anyio==4.4.0
bcrypt==4.0.1
decorator==5.1.1
fastapi==0.112.0
flake8==7.1.1
passlib==1.7.4
psycopg2==2.9.9
python-dotenv==1.0.1
pydantic==1.10.17
pydantic_core==2.20.1
PyJWT==2.9.0
pytest==8.3.2
pytest-asyncio==0.23.8
requests==2.32.3
semantic-version==2.10.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a38
sqlmodel==0.0.8
uvicorn==0.30.5
websockets==12.0
